aider: Added a new mandatory parameter `$ListName` to the `get-sp-rp.ps1` script.
# Aider chat conversation:

USER: get-sp-rp.ps1

ASSISTANT: Sure, I can help with that. Could you please provide more details about the changes you want to make to the `get-sp-rp.ps1` file?

USER: Wehn I run the script, it gives an error i.e An unknown error occurred: Cannot bind argument to parameter 'List' because it is null..

ASSISTANT: The error message "Cannot bind argument to parameter 'List' because it is null" typically occurs when you're trying to pass a null value to a parameter that doesn't accept null values. In your script, this could be happening in the following line:

```powershell
$ListItems = Get-PnPListItem -List $ListName -PageSize 2000 -Fields "FileLeafRef", "FileRef", "File_x0020_Type"
```

Here, the `-List` parameter is expecting a value, but it seems like `$ListName` is null or not defined anywhere in the script.

To fix this, you need to define `$ListName` before this line or pass a valid list name to the `-List` parameter. If `$ListName` is supposed to be a parameter of the script, you need to add it to the `param` block at the top of your script.

Here is the change you need to make:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
#Parameters
param(
    [Parameter(Mandatory=$true)]
    [string]$SiteUrl,
    [Parameter(Mandatory=$true)]
    [string]$SL,
    [Parameter(Mandatory=$false)]
    [string]$InputFile,
    [Parameter(Mandatory=$false)]
    [string]$ReportOutput = "report.csv",
    [Parameter(Mandatory=$false)]
    [string]$LogFile = "log.txt"
)
=======
#Parameters
param(
    [Parameter(Mandatory=$true)]
    [string]$SiteUrl,
    [Parameter(Mandatory=$true)]
    [string]$SL,
    [Parameter(Mandatory=$true)]
    [string]$ListName,
    [Parameter(Mandatory=$false)]
    [string]$InputFile,
    [Parameter(Mandatory=$false)]
    [string]$ReportOutput = "report.csv",
    [Parameter(Mandatory=$false)]
    [string]$LogFile = "log.txt"
)
>>>>>>> updated
```

This change adds a new mandatory parameter `$ListName` to your script. You will need to provide this parameter when you run your script.