aider: Modified `WriteProgress` function to write progress to both the log file and the console.
# Aider chat conversation:

USER: `get-sp-rp.ps1`

ASSISTANT: Sure, I can help with that. Could you please provide more details about the changes you want to make to the `get-sp-rp.ps1` file?

USER: ok so now that we have incremental option, I am no longer seeing the write progress to screen happening. Is there a way to have both happening or does what slow down the process

ASSISTANT: Yes, we can have both the incremental option and the progress display happening at the same time. The progress display should not significantly slow down the process unless the number of items is extremely large.

To achieve this, we need to modify the `WriteProgress` function in the `get-sp-rp.ps1` file to write progress to the screen as well as to the log file. We can use the `Write-Host` cmdlet to display progress on the screen.

Here is the change that needs to be made:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
# Function to write progress
function WriteProgress($Item) {
    $progressMessage = "Getting Shared Links from '$($Item.FieldValues["FileRef"])' - Processing Items $global:Counter to $($ItemCount)"
    Write-Progress -PercentComplete ($global:Counter / ($ItemCount) * 100) -Activity $progressMessage
    WriteLog $progressMessage $Item.FieldValues["FileRef"]
}
=======
# Function to write progress
function WriteProgress($Item) {
    $progressMessage = "Getting Shared Links from '$($Item.FieldValues["FileRef"])' - Processing Items $global:Counter to $($ItemCount)"
    Write-Progress -PercentComplete ($global:Counter / ($ItemCount) * 100) -Activity $progressMessage
    WriteLog $progressMessage $Item.FieldValues["FileRef"]
    Write-Host $progressMessage
}
>>>>>>> updated
```

This change will make the script write progress to both the log file and the console.