aider: Changed authentication method to use web browser login instead of credentials.
# Aider chat conversation:

USER: that does not seem to be correct as still asking for password, can this connect with web browser authenication instead

ASSISTANT: Yes, we can use the `UseWebLogin` parameter with the `Connect-PnPOnline` cmdlet to authenticate using the web browser. This will open a browser window where you can enter your credentials. Once authenticated, the script will continue to run without further interaction.

Here is the change that needs to be made:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
#Connect to PnP Online
$secpasswd = ConvertTo-SecureString "PlainTextPassword" -AsPlainText -Force
$mycreds = New-Object System.Management.Automation.PSCredential ("username", $secpasswd)
Connect-PnPOnline -Url $SiteURL -Credentials $mycreds
=======
#Connect to PnP Online
Connect-PnPOnline -Url $SiteURL -UseWebLogin
>>>>>>> updated
```

This change will make the script authenticate using the web browser. Please note that this method requires a graphical user interface to open the browser window, so it may not work if you're running the script on a headless server or a scheduled task.