aider: Updated the progress message to show "Processing Items x to total" instead of "Getting Shared Links from 'x' - Processing Items x to total".
# Aider chat conversation:

USER: the get-sp-rp.ps1 now when it is running, it shows the files been processed onscreen however the [Processing ] does not show the Processing items x to total only show this above..

ASSISTANT: I understand that you want to see the progress of the items being processed in the form of "Processing items x to total" on the screen. This can be achieved by modifying the `WriteProgress` function in the `get-sp-rp.ps1` file.

Here is the change:

```powershell
get-sp-rp.ps1
<<<<<<< HEAD
# Function to write progress
function WriteProgress($Item) {
    $progressMessage = "Getting Shared Links from '$($Item.FieldValues["FileRef"])' - Processing Items $global:Counter to $($ItemCount)"
    Write-Progress -PercentComplete ($global:Counter / ($ItemCount) * 100) -Activity $progressMessage
    WriteLog $progressMessage $Item.FieldValues["FileRef"]
    Write-Host $progressMessage
}
=======
# Function to write progress
function WriteProgress($Item) {
    $progressMessage = "Processing Items $global:Counter to $($ItemCount)"
    Write-Progress -PercentComplete ($global:Counter / ($ItemCount) * 100) -Activity $progressMessage
    WriteLog $progressMessage $Item.FieldValues["FileRef"]
    Write-Host $progressMessage
}
>>>>>>> updated
```

This change will update the progress message to show "Processing Items x to total" instead of "Getting Shared Links from 'x' - Processing Items x to total".